Wait for continueButton in KS mode if the user changed anything
This behaviour tries to minimize the "surprise factor" we have
in the installer. When the user entered spoke, changed any data
and then left the spoke, the installer gave him no chance for
review or correcting the entered values. With this patch applied
the Hub waits for a click on the continueButton if any user
interaction happened before all the spokes were ready+completed.