Fixup kickstart script logging
Changes to the execWith* methods removed the ability to handle a string
as stdin/out/err. Change the callers to first open where they want
output to be sent to and pass that instead.
Only stdout is needed because _run_program combines stdout and stderr,
pass stderr through but ignore it.

Also clean up a couple places that were passing strings that don't need
to.