Set ONBOOT=no for default autoconnections (#905918, #886090)
Autoconnections are set up by NM for devices not having ifcfg file
(ie those not activated by dracut or not configured in kickstart).
We want to default to "no" for this devices in RHEL.
We even don't want to bring up the autoconnections in installer
environment, but that is something for a separate bug.