Add more stuff to the exception reporting skip list.
For me, this reduces the time from traceback to dialog from at least ten
seconds to almost instantaneous.  And it's not helpful stuff either.  If
you see it take a very long time to generate a traceback file, please
check the file and see if it's because something is being pickled that
doesn't need to be.