Revert "Hook up the "Remove Packages" button on the dep solving error screen." (#905899).
This reverts commit f4ec3d682ffd93dfc7105eaa09acdd7fd672a3e8.

This patch also removes the button entirely.  All I've ever seen it do is reduce
the installable package set to something like 30 packages and result in really
cryptic bugs about commands that should be there failing to run.  It doesn't
seem to help in just removing the one or two packages that have a problem,
since it looks like it's frequently a core package with the problem.