Allow more than one Page to be expanded at a time.
This should help a little bit with the confusion people have where it looks
like filesystems are vanishing, when instead they are just moving to under
a different root.  However, note that anything calling _do_refresh will cause
everything to be closed except the current page.

I do not yet see a way around that.