When the user clicks "Reclaim Space", go back to the hub (#911792).
Going back to the options dialogs is arguably not really a bug (the user may
want to go to custom partitioning), but they have other ways of doing those
things.  Also, watching people test this out showed it to be very confusing
and led to people going down paths they didn't need to.  Also, it's the way
it worked in F17.