Modify repo interface in packaging
repo information is stored in 2 places, the kickstart data object, and
in the yum object (when using yumpayload). This changes the interface so
that AddOn methods reference the kickstart data and repo methods
reference the packaging data.

getRepo is now getAddOnRepo and a new getRepo has been added to
yumpayload that can be used to query the yum object.