Log when we acquire and release the _yum_lock (dlehman, clumens).
We've got a temporary locking issue somewhere that causes the UI to become
unresponsive for a bit (see #886680, perhaps other bugs).  Examining the code
indicates this isn't necessarily due to running things in the main loop
that shouldn't be.  There's something waiting for the _yum_lock.  This patch
adds logging to help debug.