Fix two places where we are locking up the main thread (#886680).
(1) Letting _first_refresh be the first time we reference payload.groups and
payload.environments means we are potentially downloading lots of big data in
the main thread, which means the UI cannot update.

(2) Referencing anything in software.py's _initialize grabs the yum lock,
possibly for a long time.  When we later check the source spoke's status, we
reference ready which checks baseRepo, which also attempts to grab the same
lock.  This isn't a deadlock because it will eventually get released, but it
does make the UI unresponsive for a while.