Add a refresh button to the custom partitioning toolbar.
This button throws away whatever changes the user has made in custom
partitioning, rescans the disks, and updates the UI to the new disk status.
This button is designed for the user who's made very specific storage
configuration changes on the shell and wants anaconda to use what they've
created.