Move custom storage setup into its own thread.
Instantiating all those device formats can result in lengthy processes like
loading kernel modules, some of which are very large.  This results in lag
between when the user leaves the welcome spoke and the hub being displayed.

In my thoroughly unscientific testing on one computer, this knocks about two
seconds off the amount of time between the welcome screen and the hub.