move Xorg test up so we might start vnc instead
Move the want_x set and check up so that, if disabled, we might still start
the VNC pathway instead. We do not need X for that path.

Also, add so more warning messages so we might know why exactly we are not
going to ask for the VNC path.