Add kickstart liveimg install command
This implements a new kickstart install command:

liveimg --url=<url> [--proxy=<proxy>] [--checksum=<sha256>] [--noverifyssl]

This will skip source and software spoke, partition the drive and then
download the image from the url location to the disk. http, https, ftp
and file methods are supported. An optional proxy can be specified,
including auth. If --checksum is passed the image checksum is confirmed
before starting to copy the files to the target system and --noverifyssl
skips checking the SSL cert when using https.

The disk images need to be able to be mounted by the installer image.
eg. extX, squashfs. If the image contains a LiveOS directory with an
image file ending in .img inside it (eg. like a live iso squashfs.img)
it will mount the inner image as the source and copy the files from
there.

eg. squashfs.img -> LiveOS/ext3fs.img -> root filesystem

Note that the system being installed *must* have the various utilities
that Anaconda expect to be installed in a normal system. This list may
change at any time, it is up to the image creator to confirm that they
have everything Anaconda needs to complete the post-install setup.