Make our gtk_* decorators safer and more intelligent
The gtk_thread_wait and gtk_thread_nowait decorators cannot be used from the
main thread as it would cause a deadlock by queueing actions to the Gtk main
loop and then blocking the loop by waiting for result to appear in the
temporary queue.

However with Thread Manager's ability to tell if we are in the main thread or
not we can make these decorators more universal so that the decorated functions
can be called from anywhere. Since they are now not only for threads, I've
renamed them to gtk_action_wait and gtk_action_nowait.

The question is if we can start decorating all methods and functions somehow
manipulating with Gtk or just methods and functions that are expected to be
called also from threads. That depends on the overhead these decorators bring.