Add geolocation module
The geolocation module exposes an API that makes it possible to get
current territory code using either GeoIP or alternatively nearby WiFi
access points. The module has multiple location backends, the default backend
is using the Fedora MirrorManager. The geolocation module is a singleton,
that is instantiated by the init_geolocation() call and refreshes geolocation info
using a thread after refresh() is called.

If the module is not instantiated and/or refreshed, None is returned by
the get_territory_code() and get_result() method.

If the module fails for some reason, it should not raise any exception,
all errors will be logged instead.