Set Tip text on the create user spoke.
Remove the 'lorem ipsum...' text and set it to:

    Keep your username less than 32 characters and do not use spaces.

I do not think it's necessary for the Tip text to explain all of the
username rules in place.  After all, if it fails, we can display more
detailed error text for the user.  The important things are length and
whether or not you can use spaces.  I'm not sure how common it is for
people to try to use symbols in their name, but I've never heard of it
and anyways, the text above is better than "lorem ipsum..."