Clean up _save_right_side and adapt to changes in blivet.devicefactory.
This changes the meaning of the encryption checkbutton to always
encrypt the leaf device. For btrfs and lvm, this means that the
behavior of encrypted autopart is different from the behavior of
add-mountpoint as they use encrypted member devices in autopart.
Unfortunately for btrfs, the only way to encrypt it is by encrypting
its members, so there is now no way to get encrypted btrfs until we
add a way to specify container-level encryption.

The new error handling in the factories also means we have to update
all of our device references after an error occurs in a factory.