Hook the Geolocation module to Anaconda
First, the module is initilized and asynchronous location info
refresh is started. Then once the welcome spoke is initialized,
it asks the geolocation module for the territory code.
If the code is unknown (either due to the lookup not returning
any result or still being in progress), default language value is used.

Also, if geolocation returns more than one language, default
langugae is used. This is because some countries, such as Switzerland,
have more than one locale and there is no clear priority which one to
use.

There is a boot option called "geoloc", that can be
used to configure the geolocation module.

It currently supports these values:
not set - use geolocation with default provider
"0" - don't use geolocation
"provider_mirror_manager" - use the Fedora MirrorManager as the
                          geolocation provider (default)
"provider_hostip" - use Hostip.info as geolocation provider

The "provider_*" strings are defined by the "GEOLOC_PROVIDER_*"
constants in the constants module.