Add module providing safe DBus operations
We had a lot of troubles with the DBus and the way we use threads and
main loops. Our is_wireless_device defined in isys and written in C caused
segfaults. It helped to rewrite it in Python. Now it is interaction with
systemd-localed over DBus that causes segfaults in spite of the fact that
it is a Python code. Modifying it to use Gio's sync proxy instead of Python's
dbus module (that uses probably not thread-safe libdbus) didn't help as the
proxy uses the concept of cached properties that are probably updated by
signals emitted by the object and thus not available synchronously and maybe
not updated at all in some cases because our main loop runs in a different
thread than the thread we want to use the functionality.

So I guess it is time for our own DBus module that uses Gio's primitives to
provide both thread and main loop safe DBus operations.

Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>