Fix console= persistence, remove serial (#928269)
The 'serial' option was supposed to have been removed, but it got left
behind in a few places. Somehow it mutated into an argument to keep the
user-set "console=XXX" argument in the installed system?

This patch should straighten things out to match the intended behavior:
* "serial" is deprecated and ignored; try "console=ttyS0" instead.
* If you use "console=XXX" it will be preserved in the installed system.
* If multiple "console=XXX" args are used, the *last* one is preserved.

It might be kind of useful to be able to override this but I really
can't think of a use case for that right now, so I'm leaving it out.
You can always just remove it with a sed oneliner¹ or something.

Here's what the patch actually changes:
- Adds iutil.get_active_console()
- Uses that to fix iutil.isConsoleOnVirtualTerminal()²
- Sets BootLoader.console whenever 'console=XXX' is present, instead of
  only when flags.serial is set
- Uses iutil.isConsoleOnVirtualTerminal() instead of flags.serial to
  determine when it's OK to use GRUB splash
- Removes flags.serial, since it's now completely unused.

¹ e.g. `sed 's/console=\S*\s*//' /etc/grub*.cfg`
² Note the "XXX PJFIX", which means pjones owes me a beer.