Move text UI summary hub setup into the setup method (#927315, #950956).
Without this, self._keys was empty (because it's set up by the superclass's
setup method, which gets called after the summary hub's __init__).  Thus,
there were no spokes and no spoke execute methods got called.  Thus, the
storage configuration in the kickstart file was not applied and we had to
stop and ask.