Disable the "Closest mirror" option if there's no fastestmirror plugin (#876135).
The goal here is to not allow selecting the Closest mirror option on trees that
do not include a repo config file supporting it.  This seems to be the easiest
way to do it, though I have not been able to test it thoroughly due to trees
being broken in other ways.  It does not break the standard Fedora case, though.

Additionally, I've moved Closest mirror to the last option in the combo box so
it can be removed without having to move all the other options around.