Handle both types of data we can get from libxklavier (#950921)
Since we have to work with multiple versions of libxklavier (Fedora, RHEL, ...)
that may be built with various versions of introspection-generating tools, we
should both types of data we can get.

I still hope the change that caused we get list of integers instead of a string
will be reverted one day.