Make several changes to how addons are displayed (#873498).
(1) Change the labels at the top of the software spoke so it looks like the
thing on the left side and the thing on the right side are related.

(2) Show environment-specific addons at the top of the list, followed by
generic ones.  Each set of addons is sorted alphabetically.

(3) Display a thin separator after the environment-specific addons.