Add Driver Update Disk support to anaconda-dracut
This hooks into dracut in order to detect disks labeled with OEMDRV or
use specified devices to examine for driver update disk repositories.
Drivers can be selected interactivly by passing inst.dd on the cmdline.
The user will be presented with a list of driver rpms from the
repositories that it discovers on the mounted devices.

Repositories that are used to install a driver are copied over to
/run/install/DD-X directories for use by Anaconda when installing the
system. The list of selected drivers is written to
/run/install/dd_packages

Drivers can also be retrieved from the network with inst.dd=http|ftp|nfs
by they will not be installed to the target system because they are a
single rpm, not a yum repository.