Move the selected language to the top of the list on the welcome screen.
This is how it was mocked up to begin with.  Also, people don't like getting
dumped into the middle of a list - it's better if they start at the top and
can only scroll down.  And finally, doing this gets rid of the problems we've
occassionally seen where the chosen language is either completely or partially
off the top of the view.