Add support for the realm command
This patch adds support for the realm command directly to Anaconda,
so that realm can be specified directly with the realm command without
the addon data notation.

The realm command enables joining a domain during installation.
Only the join command is currently supported.

This patch is based on the original Stef's patch and the Vratislav's realm
addon.