Streamline DatetimeSpoke's timezone updating (#953311)
Without this patch the final word about the selected timezone was on the
TimezoneMap which brought bugs and complicated things because the map doesn't store
the timezone it was asked to select and when queried, it returns the value it
calculates from the currently selected timezone. Thus when selecting e.g.
US/Central, the map selected it correctly (visually), but when queried with
get_timezone afterwards it returned America/Chicago.

All in all we shouldn't rely on the map having the best data when it does some
transformations on its own and does not necessary support all timezones we want
to set (now or in the future).