Fix a probably rare traceback in resizing from the custom part UI.
As far as I can tell, this really only happens if you have previous partitions
taking up the entire disk, blank out the desired capacity for a partition that
is resizeable, and click the update button.  Since it cannot be resized, we
hit a traceback that our earlier tests for this condition do not catch.