Fix non-default language being hidden in welcome spoke
Currently if Geolocation supplies a language other than the default,
it is correctly added to the top of the list in welcome spoke, but it
is outside of the viewport and thus not visible.
Using scroll_to_cell with use_align=True and row_align=0.5 should fix this.
Also, _selectLanguage has been modified to correctly use the filter
data model every time (previously it was sometimes using the treeview
data model directly and sometimes the filter).