Relabel the disk when using autopart all (#758881)
When autopart all is selected and all partitions have been removed from
a disk go ahead and apply the correct disk label. This fixed the problem
of running autopart all on a gpt disk in a non-EFI system. Also add a
few more lines to the logfile when relabeling the disk.