dracut: minor shell cleanup
getargbool doesn't echo any result, so $(getargbool ...) will always
evaluate to the empty string (""). This works fine in this case:

    if $(getargbool ...); then ...

because the 'if' statement uses the *return* value of the command, not its
output. But if getargbool echoed a result (like getarg), this would
break. So remove the subshell escape.