Add a "Delete All" button to the reclaim dialog.
This allows for clearing all editable disks in one button click.  It is
positioned on the far right of the list of action buttons and schedules a
delete action for all disks.  Should you later wish to undo this action,
the button will also do a preserve all.

This also required moving the shrink slider down.