Use the new Fedora GeoIP API (#963531)
Remove the old MirroManager geolocation provider and replace it with
Fedora GeoIP provider based on the new API. The API directly provides
the time zone and appears to be much more precise than the old one.
If no time zone is returned, Anaconda will fall back to getting the
time zone from the territory code.