Convert escaped spaces on cmdline to _ (#874381)
We use shlex.split to parse the cmdline. This mostly works, except that
it removes any \\ from the string instead of converting them back to \

This results in the device LABEL having x20 in it, which isn't matching
the ID_FS_LABEL when looking for protected devices.

So, this replaces the \\x20 with _ when parsing the cmdline, which also
matches what ID_FS_LABEL expects. This results in the usb device being
filtered from the available disks for the install.