Use inline completion for the region/city selection
Due to the still unresolved Gtk bug #907946, the completion for regions and
cities doesn't work properly. As it seem nobody will resolve the bug soon, I'd
like to use inline completion in the meantime. It doesn't provide such good
functionality as it appears only when there is unique match, but it's better
than nothing.