Update transifex docs
The ordering was wrong, you need to change the config after pulling from
master, otherwise it tries to pull from the new branch which is empty.

You also don't need to create the new resource with the transifex UI,
using the tx set is sufficient. tx push will create the new resource for
us.