Use install tree from repo=hd (#818233)
This adds the ability to use an install tree with repo=hd:... which is
helpful for dd'd DVD iso's on USB drives. The first bit allows storage
to be initialized when a USB device has been skipped because it doesn't
have partitions, just an iso9660 image.

The other part checks the hd: path to see if repodata/repomd.xml exists,
and if so sets that up as the baserepo. Otherwise it falls back to
looking for iso images in the directory as it would normally.

Resolves: rhbz#818233