Check if selected repos have the core package group (#901515)
Check if the selected repositories contain the "core" package group,
which is a requirement for successful installation.
In GUI:
If the core group is not available, show warning dialog if next is clicked.
If core is available, let the installation proceed as usual.
In TUI:
If the core group is not available, show warning dialog and
exit once OK is clicked.