Remove the cryptic "language-default keyboard" checkbutton
People have problems to understand what that checkbutton does and expect the
language-default layout to be set up and activated by default. Now, that we have
the layout indicator, it should be okay to do so as users can see the change of
the current layout. For language-default layouts that do not support typing
ASCII characters, we should append 'us' and activate layout switching.