Rework the Welcome spoke to allow users choose from all locales
What we actually need from users is a locale, value that can be set in the $LANG
environment variable. However, what we get from the files with translations are
usually languages (en, cs, es, de, ...). Instead of picking up one locale for
each such language, we may offer user all locales for all languages we have
translations for and let them choose the exact locale they want to use. By
setting it as the value of the $LANG variable, we then get the right
translations.

Resolves: rhbz#951879
Related: rhbz#986155