Use the active window as parent for new dialogs (#888292)
When creating a new window in the GUI InstallInterface, use the current
active window as the transient parent instead of always using the
InstallControlWindow instance. For most windows this is effectively the
same window, but for dialogs spawned from other dialogs, this allows the
control (and correct window focus) to return to the spawning window when
a dialog is finished.

If no active window can be found, the InstallControlWindow is used as
before.