Fix handling of non-ASCII names (#968470)
Applied 65a72fd from master (#969309):

When converting a full name into a username, decode the entire first
word of the name before taking the first initial. Before the string
is decoded, the first byte may be the first byte of a UTF-8 encoded code
point instead of a complete character.