Do not populate devicetree after each single login in iscsi dialog (#740105)
Instead, populate it when leaving the dialog.
Populating the tree after each single login is not necessary and it is causing
this issue in iface-bound iSCSI multipath scenario (eg sdf being a node bound
via p3p1 and sdg being the node bound via p3p2):

The first populate after logging via p3p1 sets a format (eg DiskLabel) for sdf.
In the second populate after logging via p3p2 creating sdg, the sdg format is
correctly identified as multipath_member, but for sdf original format is found
and kept.  As a consequence, sdf is not filtered out by
pyanaconda/ui/lib/disks.py:GetDisks() function and is displayed in filter UI
(as well as in advanced storage disk overview).

With the patch, the filter UI correctly displays only mpath device (as in
RHEL6) - if both multipath members / iscsi sessions are added in single add
iscsi dialog run instance, which is still improvement.