Don't show multipath members in specialized disks overview (#740105)
With previous patch, the members shouldn't be returned by
pyanaconda/ui/lib/disks.py:getDisks(), but because they are added to
data.ignoredisk.onlyuse as ancestors in apply of filter spoke, they are still
displayed in the overview. I am not sure if we really need to do the ancestors
adding so not to break something I just filter the onlyuse disks by disks
returned by getDisks() in this patch.