Fix crash while parsing ntp servers from DHCP6 (#987978)
This removes anaconda's attempt to collect NTP servers through
NetworkManager's DHCP6 options. The reasons for this are as follows:

  - NetworkManager does not currently request NTP or SNTP options. When
    using a DHCP6 that honors OPTION_ORO (e.g., ISC dhcpd6), no time
    servers will be returned
  - dhclient does not support the NTP server option (dhcp6 option 56,
    RFC 5908). If a server does return an OPTION_NTP_SERVER block, it
    will not be parsed and NetworkManager will only be aware of the
    presence of an unknown option 56.

It is possible to use the OPTION_SNTP_SERVER data, which is formated in
the same way the IPv4 DHCP NTP server option data. To use this we would
need to create /etc/NetworkManager/dhclient6.conf containing "also
request dhcp6.sntp-servers" and fix the option parsing code in nm.py to
pull the right value out of the Dhcp6Config.Options dictionary. However,
it's probably not worth it, since obviously no one was using DHCP6
options at all, because it would have crashed.