Enable the AMD IOMMU driver if tboot is installed. (#928144)
If tboot is installed, enable the AMD iommu driver in the kernel
parameters in the same way that we enable the Intel iommu driver.

The AMD iommu driver is disabled by default in RHEL 6 due to stability
issues with some BIOSes after suspend and resume. This kernel parameter
has no effect on systems without AMD-Vi.