Add automatic fetching of RPMs for new Defines & Requires
When the -f or --fetch flag is used, the Anaconda spec file
diff will be checked for new packages specified with Defines and
Requires and will then automatically download and include RPMs
belonging to those packages in the updates image.
The RPMs are cached locally (in the ~/.anaconda_updates_rpm_cache folder)
and are not downloaded again on the next run.
Architecture selection is supported by specifying the architecture
(for example i686 or x86_64) as an argument to the -f flag.

Signed-off-by: Martin Kolman <mkolman@redhat.com>