Handle focus changes of MountpointSelectors from outside (#975838)
We handle clicking and pressing keys on the MountpointSelectors from outside
which allows us to do some magic the selectors have no clue about. Doing the
same with focus changes (that happen e.g. when using arrows) allows us to keep
the MountpointSelector selected (with blue background and the right arrow shown)
even if the focus moves to some other part of the window.