Set system date and time with our own function
Now that we have a function to set system time, we could wrap it with one more
and use it instead of running 'date -s' which could be problematic and prints
out the time and date to tty1.

Also add logging to the set_system_time function, so that we can see why the
timestamps suddenly changed in the logs.