Add unsupported hardware dialog (#872728)
In GUI installs this adds a dialog after the Welcome screen that warns
if the hardware is unsupported. The dialog is skipped only if
unsupported_hardware is included in a kickstart file.

In text installs this adds a standalone spoke before the first hub that
displays the warning message.

(cherry picked from commit fff0eb74f1c8dc81fcc0844aacf077e25b695d33)