Add support for network configuration in TUI. (#909299)
After re-writing it about three times, I think this is as non-ugly as
the TUI network spoke is going to get.

There are a couple of exceptionally ugly hacks in here. I wanted to use
nmcli to be able to add/configure devices, but that functionality is not
present--hence the need for the ugly hacks. nmcli is, however, under
active development to add these capabilities, so hopefully it is not too
long before a new NM is released with a more robust nmcli. When that
happens, large swaths of this can be re-written much more cleanly.

That said, only wired connections are supported in this. Support for
vlan/bonding/wifi devices would add far too much complexity for what
should be a clean and simple interface--at least at this point. So,
users need to install graphically or via kickstart if they want to
configure those devices.

Also, I hope Radek won't object to my including his name on the authors
list. I copied over a non-trivial portion of the back-end logic, so I
feel it should get added.