Don't wait for systemctl shutdown command to exit (#974383)
Using os.system to wait for shutdown means it blocks, we get a SIGTERM
but can't act on it so it delays for 90 seconds and sends SIGKILL,
making the shutdown slow.

Use subprocess.Popen instead so we can exit while systemctl does the
shutdown.

Also don't use sys.exit in forked child processes, they don't need to
call the exit handler, the parent will handle that.