Fix all the "X is defined outside of __init__" warnings.
There were an awful lot of these, and I took a couple tactics when fixing them.
The first is that in a lot of places, we define attributes before they are
ever used, which is what pylint is so concerned about here.  To ignore those,
I added a pylint command line argument.  Another tactic was to move a lot of
the get_object calls into _grabObjects or __init__ or somewhere.  I was also
able to remove some unused variables.