Fix up all unused variable warnings.
This is going to seem like a really picky commit.  In a lot of places, I've
changed the code to just grab one element out of a tuple.  In other places,
I've prefixed the name of an unused variable with an underscore.  pylint will
ignore these, and it makes for clearer code than trying to do something
else.

Even though it's picky, this test is worth keeping around because it does find
real errors.