Handle >=,<=,= for package version, fix -a/--add
Handle correctly the >=, <= and = operators
used in Requires and get a matching package
version from Koji.
NOTE: The newest matching package version
from Koji is used.
Also don't include packages specified in defines,
if they are not specified in Requires.

Fix handling for RPMs added by -a or -add, print all added RPMs
 and process all paths with os.path.abspath(), so that both
relative and absolute paths can be safely used.
Correctly normalize the list of all added paths, remove duplicates
and skip "manually" added packages during autofetch.

Please note that if a package is added using the -a/--add option,
it will override any other packages that might be available either
in Koji or in the local cache and any version mismatch will be _ignored_.
The version is ignored on purpose, so that individual packages can
be manually overridden if needed.