Do not run another instance of the TUI for errors (#997661)
Running another instance of the TUI possibly from a non-main thread may result
in a weird and unexpected behaviour. Instead of running a new TUI, we may use
our messaging system to request dialogs to be run in the main thread.

This patch basically does the same as the @gtk_action_wait decorator, just
replacing the missing equivalent of GLib.idle_add with message passing and
processing.