Let Gtk pick the size for the isoChooserDialog
If it's not big enough now, it's not our problem.

Also converted the dialog from a GtkDialog containing a
GtkFileChooserWidget into just a GtkFileChooserDialog, which is
basically the same thing but one less layer of objects that we have to
define for ourselves.

Unrelated things that glade changed but that seem ok enough to keep: a
<style> element was moved from the middle of an <object> element to the
end of it, and two instances of "icon-size" were changed to "icon_size".

Resolves: rhbz#1063840

(merged from commits 941d9e7ea843dae8cb51511b3096fa18bb607455 and
 17214635b4faa54f3d529f642897f6e16bcfc6f8)