makebumpver: Fix problem with single line body
If the body of a commit was a single line it was returned as a string
when a list is expected. This could cause it to think there was no bug
referenced. Check for this and make it always a list.

Also cleanup some variable names which were overriding builtins or
global variables.