Fix a SIGSEGV when returning from storage spoke (#983319)
This is kinda gross. Backstory:

 - As part of creating a new lightbox (anaconda_lb_show_over), we attach
   a signal handler (anaconda_lb_move_window_to_parent) to the
   configure-event signal on the window contained in the lightbox. One
   of the parameters to this signal handler is the lightbox itself.
 - Normally this signal handler is called as soon as enlightbox yields
   to the caller, because that's when the widget is displayed and thus
   when the first size/position/stacking change occurs. The signals stop
   when the lightbox is destroyed, because the window can't change
   anymore.
 - Except this isn't what happened. When returning from the storage
   spoke, a new enlightbox dialog was created, and then destroyed, and
   then the signal handler (that takes the (destroyed) lightbox as a
   parameter) was called.
 - Kaboom.

The solution here is to disconnect the signal handler from the window as
the containing lightbox is destroyed, and also go ahead and emit the
signal if it never fired in the first place because otherwise the
stacking is going to be all messed up. This is not a very good solution!
But hopefully it'll work until we can figure what went wrong with
setting up the lightbox in the first place.