Add new logger for sensitive information
Content that goes to the sensitive-info logger is not forwarded to any
other log and the logfile - sensitive-info.log - is not copied to the
installed system once installation is finished.

Like this, potentially sensitive information is available for debugging
once things go wrong during the installation, but only it the user
willingly decides to provide it.

Signed-off-by: Martin Kolman <mkolman@redhat.com>