Get rid of the non-deterministic expand_langs and its usage
Using expand_langs was non-deterministic in a way that if e.g. "pt" came before
"pt_BR" in searching for a match for something from expand_langs("pt_BR"), it
was picked up. However, in such a case we want to pick up "pt_BR" as it matches
better.

This patch adds a function for finding matches (now the best one)
deterministicly and a function to determine if given langcode (e.g. "en")
matches some locale (e.g. "en_US"). The former one is useful e.g. for getting
the best rnotes to show based on the installation language, the latter one is
useful for adding langsupport groups to the installation transaction.