Run firstboot-only spokes on first boot by default
Spokes that should appear in both the installer and the initial-setup during the
first boot should run only if 'firstboot --reconfig' is specified in the
kickstart. But firstboot-only spokes should always run by default during the
first boot, because otherwise they wouldn't be shown at all. The should_run
classmethod can be overriden if needed, of course.

Related: rhbz#1000409