Let users configure keyboard via anaconda in live installations (#1002533)
Setting up keyboard layouts for both the installation process and the installed
system is a complicated and confusing task. Because the default environments
don't show the layout indicator and have keyboard configuration hidden in some
menus, we may let users configure keyboard layouts via anaconda.

If somebody configures keyboard on their own and the anaconda's configuration
for the installed system is different, we should warn them and make them visit
the Keyboard spoke to confirm the configuration.