Don't mark spoke as completed if no repo is set. (#1001538)
Even though the status text notifies users there is a problem
setting up the install repo, the spoke should not be getting
detected as 'complete'.

In the GUI, this now marks the spoke with the error icon so it's
more visibly obvious to users the spoke needs attention.

In the TUI, the '!' (for 'not complete') is shown next to the spoke
instead of 'x' ('complete').

Resolves: rhbz#1001538