Remove magic from the passphrase dialog (#921948)
Playing with sensitivity and jumping between widgets based on equality of the
password and confirmation entries is confusing and may cause problems. The
confirmation entry cannot be clicked when the password entry is being eddited,
only TAB works for switching into it. And jumping to the save button when the
confirmation entry equals the password entry may lead in an incomplete password
entered and "confirmed".

Hitting ENTER in one of the entries may click the Save button if both entries
have the same non-empty string.