Move the intltool Makefile rules into configure.ac
This copies the rule for running intltool-merge on .desktop files from
IT_PROG_INTLTOOL in intltool.m4. We're not using the macro itself
because it expects us to run intltoolize on the source tree, and the
intltool version of Makefile.in.in breaks some of the things gettext
does. But at least this way the intltool-merge rule is in one place
instead of two, and it works right when run outside of $srcdir.

Also added autoconf checks for the intltool programs.

Related: rhbz#1121678