Fix handling of flexible specs in onpart for member devices. (#1004885)
Short version: don't assume onPart identifier is a name/path.

Longer version:
If an existing pv was specified using --onpart=UUID=<uuid> the
immediate onpart resolution worked, but the volgroup would then fail
to resolve the pv.xxx to a device because it assumed the onpart data
was a name.