Don't use temporary file and move when writing out an ifcfg file
NetworkManager uses IN_CLOSE_WRITE inotify event which is not triggered if a new
file is moved to the place of the watched file. Also the SELinux context is
wrong when a file from /tmp is used to the /etc tree.

Related: rhbz#972362

Based on the patch from Hans de Goede <hdegoede@redhat.com>.