Fix pre-processing of files for xgettext.
This combines the Rules-glade and Rules-liveinst into one file and
modifies the rules so that they work with all of our various glade and
liveinst paths. Update POTFILES.in to point to the generated files so
that xgettext doesn't try to process the source files directly.

For files that xgettext can handle but can't autodetect the type,
manually extract strings into a .po file. xgettext can import .po files
into the .pot and knows to leave the file names and line numbers alone.

Related: rhbz#1121678