Don't set ksdata.lang.seen to True if using default value (#997397)
If setting kdsdata.lang.lang to the default value (DEFAULT_LANG),
don't set ksdata.lang.seen to True. Otherwise it would not be possible
to discern the default value from a value set by the user using
kickstart or a bootoption.

Also clean-up the language handling code a bit and remove
the default value setting in kickstart.py as it is no longer needed.